[ISSUE-1]: Rename application to Composed Contacts - Regenerate android and ios projects - Relink native packages - Update react-native-navigation to latest
