+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.pumpapp",
+    package = "com.composedcontacts",
 )
 
 android_resource(
     name = "res",
-    package = "com.pumpapp",
+    package = "com.composedcontacts",
     res = "src/main/res",
 )
