+++ android/app/build.gradle
@@ -33,6 +33,13 @@ import com.android.build.OutputFile
  *   // bundleInPaidRelease: true,
  *   // bundleInBeta: true,
  *
+ *   // whether to disable dev mode in custom build variants (by default only disabled in release)
+ *   // for example: to disable dev mode in the staging build type (if configured)
+ *   devDisabledInStaging: true,
+ *   // The configuration property can be in the following formats
+ *   //         'devDisabledIn${productFlavor}${buildType}'
+ *   //         'devDisabledIn${buildType}'
+ *
  *   // the root of your project, i.e. where "package.json" lives
  *   root: "../../",
  *
@@ -65,6 +72,10 @@ import com.android.build.OutputFile
  * ]
  */
 
+project.ext.react = [
+    entryFile: "index.js"
+]
+
 apply from: "../../node_modules/react-native/react.gradle"
 
 /**
@@ -87,7 +98,7 @@ android {
     buildToolsVersion "25.0.1"
 
     defaultConfig {
-        applicationId "com.reactnativestarter"
+        applicationId "com.composedcontacts"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
@@ -126,13 +137,13 @@ android {
 }
 
 dependencies {
-compile project(':react-native-linear-gradient')
-compile project(':react-native-vector-icons')
-compile project(':react-native-navigation')
-compile project(':react-native-locale-detector')
+    compile project(':react-native-vector-icons')
+    compile project(':react-native-locale-detector')
+    compile project(':react-native-linear-gradient')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
+    compile project(':react-native-navigation')
 }
 
 // Run this once to be able to run the application with BUCK
