+++ android/app/src/main/java/com/composedcontacts/MainActivity.java
@@ -1,0 +1,5 @@
+package com.composedcontacts;
+
+import com.reactnativenavigation.controllers.SplashActivity;
+
+ public class MainActivity extends SplashActivity {}
