--- android/app/src/main/java/com/reactnativestarter/MainActivity.java
@@ -1,5 +1,0 @@
-package com.reactnativestarter;
-
-import com.reactnativenavigation.controllers.SplashActivity;
-
-public class MainActivity extends SplashActivity {}
