--- android/app/src/main/java/com/reactnativestarter/MainApplication.java
@@ -1,38 +1,0 @@
-package com.reactnativestarter;
-
-import android.app.Application;
-
-import com.facebook.react.ReactApplication;
-import com.BV.LinearGradient.LinearGradientPackage;
-import com.oblador.vectoricons.VectorIconsPackage;
-import com.reactnativenavigation.NavigationApplication;
-import com.i18n.reactnativei18n.ReactNativeI18n;
-import com.facebook.react.ReactNativeHost;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-import com.facebook.soloader.SoLoader;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class MainApplication extends NavigationApplication {
-
-  @Override
-  public boolean isDebug() {
-     // Make sure you are using BuildConfig from your own application
-     return BuildConfig.DEBUG;
-  }
-
-  protected List<ReactPackage> getPackages() {
-    return Arrays.<ReactPackage>asList(
-      new VectorIconsPackage(),
-      new ReactNativeI18n(),
-      new LinearGradientPackage()
-    );
-  }
-
-  @Override
-  public List<ReactPackage> createAdditionalReactPackages() {
-     return getPackages();
-  }
-}
