+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ReactNativeStarter",
-  "displayName": "React Native Starter"
+  "name": "ComposedContacts",
+  "displayName": "Composed Contacts"
 }
