--- index.ios.js
@@ -1,1 +1,0 @@
-import './src/index';
