+++ ios/ComposedContacts/AppDelegate.h
