+++ ios/ComposedContacts/Images.xcassets/AppIcon.appiconset/Contents.json
