+++ ios/ComposedContacts/Images.xcassets/Contents.json
