+++ ios/ComposedContactsTests/Info.plist
