+++ package.json
@@ -22,7 +22,7 @@
     "react-native": "^0.50.1",
     "react-native-linear-gradient": "^2.3.0",
     "react-native-locale-detector": "^1.0.1",
-    "react-native-navigation": "^1.1.256",
+    "react-native-navigation": "^1.1.336",
     "react-native-style-tachyons": "^3.4.2",
     "react-native-vector-icons": "^4.3.0",
     "react-redux": "^5.0.5",
